projects/excel-addin/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic-settings
sqlalchemy[asyncio]
asyncpg
alembic
pyjwt
cryptography
